cpu affinity: consider --cpu-offset as override to auto-detection
The --cpu-offset arugment was incompatible with automated detection of
free/busy CPUs. The reduced cpu set is not obvious during startup time
and we may fail to launch workers if cpu set + cpu offset < workers.

This patch disables auto-detection of free vcpus if a --cpu-offset
is given. This restores the original semantics of --cpu-offset.